backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
torch==2.2.1
torchvision==0.17.1
matplotlib==3.8.3
Pillow==10.2.0
python-json-logger==2.0.7
structlog==24.1.0
transformers==4.38.2 